fishished GUI
Fixed bug where yt-dlp would skip downloads causing the GUI to look like its loading indefinitely

Fixed bug with file names on the download page where they would be truncated incorrectally

Added finish button to enable when download is finished and text to say the download is finished
